requirements.txt
# DocLayNet Model Inference Requirements
# =====================================

# Core ML/DL frameworks
torch>=1.12.0
torchvision>=0.13.0
torch-geometric>=2.1.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Image processing
Pillow>=9.0.0
opencv-python-headless>=4.5.0


# Data processing
scikit-learn>=1.1.0
matplotlib>=3.5.0

# Utilities
requests>=2.28.0
tqdm>=4.64.0



# Data handling + visualization
matplotlib
scikit-learn

# PDF + image processing
pdf2image
# Optional (if used elsewhere)
networkx
